decode: mark destroyed buffers as deleted.
In multi-threaded scenarios destroying a buffer via destroy_va_buffer()
will allow other threads to claim the freed buffer object. If the
calling code does not know whether the buffer object was freed it may
attempt to vaDestroyBuffer, essentially destroying a buffer object which
is in use by another thread.